Split `type_of` out of collect.rs
